+++ src/main/java/com/reggiemcdonald/neural/net/OutputNeuron.java
@@ -62,7 +62,7 @@ public class OutputNeuron implements Neuron {
     }
 
     @Override
-    public void setBias(float bias) {
+    public void setBias(double bias) {
         this.bias = bias;
     }
